# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resolve API",
    "version": "1.0.0-beta",
    "description": "Automated resolution infrastructure for long-tail prediction markets. Verdicts are contract-validated; UNRESOLVED always carries a caveat; a non-observation is never a negative verdict. Percentages in the track record appear only after 100 reconciled markets per platform."
  },
  "servers": [
    {
      "url": "https://resolve.rafaemush.workers.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "MarketRegistration": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "platform": {
            "default": "custom",
            "type": "string",
            "enum": [
              "polymarket",
              "limitless",
              "custom"
            ]
          },
          "external_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "condition": {
            "type": "string",
            "minLength": 8,
            "maxLength": 4000
          },
          "event_statement": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "option_a": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "option_b": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "positive_option": {
            "type": "string",
            "enum": [
              "OPTION_A",
              "OPTION_B"
            ]
          },
          "anchors": {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200
            }
          },
          "sources": {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "kind": {
                  "type": "string",
                  "enum": [
                    "github_api",
                    "github_events",
                    "base_log",
                    "solana_log",
                    "web_fetch",
                    "web_render"
                  ]
                },
                "ref": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 2048
                }
              },
              "required": [
                "kind",
                "ref"
              ],
              "additionalProperties": false
            }
          },
          "open_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "deadline_utc": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "grace_seconds": {
            "default": 3600,
            "type": "integer",
            "minimum": 0,
            "maximum": 2592000
          },
          "resolver": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "github_pr_merged"
                  },
                  "repo": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_.-]+\\/[A-Za-z0-9_.-]+$"
                  },
                  "pr": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "kind",
                  "repo",
                  "pr"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "github_release_published"
                  },
                  "repo": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_.-]+\\/[A-Za-z0-9_.-]+$"
                  },
                  "tag": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "kind",
                  "repo",
                  "tag"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "github_issue_closed"
                  },
                  "repo": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_.-]+\\/[A-Za-z0-9_.-]+$"
                  },
                  "issue": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "kind",
                  "repo",
                  "issue"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "evm_log_present"
                  },
                  "chain": {
                    "type": "string",
                    "const": "base"
                  },
                  "address": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]+$"
                  },
                  "topic0": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]+$"
                  },
                  "topics": {
                    "maxItems": 3,
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^0x[0-9a-fA-F]+$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                },
                "required": [
                  "kind",
                  "chain",
                  "address",
                  "topic0"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "solana_sig_present"
                  },
                  "account": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 64
                  },
                  "discriminator": {
                    "type": "string"
                  }
                },
                "required": [
                  "kind",
                  "account"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "const": "numeric_threshold"
                  },
                  "path": {
                    "type": "string",
                    "minLength": 1
                  },
                  "op": {
                    "type": "string",
                    "enum": [
                      ">=",
                      "<=",
                      ">",
                      "<",
                      "=="
                    ]
                  },
                  "value": {
                    "type": "number"
                  },
                  "unit": {
                    "type": "string"
                  }
                },
                "required": [
                  "kind",
                  "path",
                  "op",
                  "value"
                ],
                "additionalProperties": false
              }
            ]
          },
          "negative_rule": {
            "default": "absence_after_deadline",
            "type": "string",
            "enum": [
              "absence_after_deadline",
              "explicit_negative"
            ]
          },
          "allow_prerelease": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "platform",
          "external_id",
          "condition",
          "event_statement",
          "option_a",
          "option_b",
          "positive_option",
          "anchors",
          "sources",
          "open_at",
          "deadline_utc",
          "grace_seconds",
          "negative_rule",
          "allow_prerelease"
        ],
        "additionalProperties": false
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data",
          "request_id",
          "schema_version"
        ],
        "properties": {
          "ok": {
            "const": true
          },
          "data": {},
          "request_id": {
            "type": "string"
          },
          "schema_version": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "error",
          "request_id",
          "schema_version"
        ],
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          },
          "request_id": {
            "type": "string"
          },
          "schema_version": {
            "type": "string"
          },
          "remaining": {
            "type": "integer"
          },
          "resets_at": {
            "type": "string"
          },
          "retry_after_seconds": {
            "type": "integer"
          },
          "how_to_proceed": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/markets": {
      "post": {
        "summary": "Register a market and its watches (robots and contract checks at registration)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "market_id, status (open | unsupported_source), reasons, watches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "validation_error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "watch limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List markets",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "markets[]",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{id}": {
      "get": {
        "summary": "Market with watches",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "market, watches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a market and deactivate its watches",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/webhooks": {
      "post": {
        "summary": "Register an https endpoint (secret shown once). events: a subset of market.resolved, market.unresolved_update, market.error, credits.low, payment.credited, shadow.committed, shadow.revealed (default: all). The first delivery attempt is made as the event happens (a few per event; the rest by the 5-minute queue drain); failed attempts retry on a backoff of 60 s, 5 min, 30 min, 2 h, 12 h, 24 h, then dead-letter",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "pattern": "^https://"
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "market.resolved",
                        "market.unresolved_update",
                        "market.error",
                        "credits.low",
                        "payment.credited",
                        "shadow.committed",
                        "shadow.revealed"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "endpoint + secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "validation_error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List endpoints",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "endpoints[]",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/webhooks/{id}": {
      "delete": {
        "summary": "Deactivate an endpoint",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}