# src/api/internal.ts
import { Hono } from "hono";
import type { Env } from "../env";
import { parseConfig } from "../env";
import { ok, err, waitUntilOf } from "./envelope";
import { safeEqual, bearer } from "./admin";
import { hmacHex } from "../resolve/text";
import { runWatch } from "../ingest/watch";
import { registerMarket } from "../markets/register";
import { db, rpc } from "../db/supabase";
import { makeJevCaller } from "../jev/client";
import { mintKey } from "./keys";
import { runReconcile } from "../jobs/reconcile";
import { scanDeposits } from "../jobs/deposits";
import { drainWebhooks } from "../webhooks/deliver";
import { Budget, INVOCATION_SUBREQUESTS } from "../ops/budget";

type Vars = { requestId: string; schemaVersion: string };
export const internal = new Hono<{ Bindings: Env; Variables: Vars }>();

const isAdmin = (c: { req: { header: (n: string) => string | undefined }; env: Env }) => { const k = bearer(c as never); return !!k && safeEqual(k, c.env.ADMIN_API_KEY); };

/** pg_net -> one watch poll. Signature = HMAC(secret, "<watch_id>|<YYYY-MM-DDTHH:MM>") over the dispatch minute (+-3 min tolerance). Admin bearer also accepted for manual runs. */
internal.post("/watch/:id", async (c) => {
  const id = c.req.param("id");
  if (!isAdmin(c)) {
    const sig = c.req.header("x-internal-signature") ?? "";
    const minute = c.req.header("x-internal-minute") ?? "";
    const t = Date.parse(minute + ":00Z");
    if (!sig || !Number.isFinite(t) || Math.abs(Date.now() - t) > 3 * 60_000) return err(c, "forbidden", "bad or stale internal signature", 403);
    const expected = await hmacHex(c.env.INTERNAL_HMAC_SECRET, `${id}|${minute}`);
    if (!safeEqual(sig, expected)) return err(c, "forbidden", "invalid internal signature", 403);
  }
  const cfg = parseConfig(c.env);
  const s = await runWatch(c.env, cfg, id, { waitUntil: waitUntilOf(c) });
  // pg_net stores this status in net._http_response and dispatch_failures() (migration 013) counts >= 400 as a poll that
  // did not happen. A run that recorded its outcome, 'failure' included (a source error, alerted by the runner's own
  // transition and streak logic), is a delivered dispatch; only a run that could not record itself (loop_runs row or
  // watch bookkeeping) answers 500.
  return ok(c, s, s.recorded === false ? 500 : 200);
});

internal.post("/markets", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const cfg = parseConfig(c.env);
  const body = (await c.req.json().catch(() => null)) as { market?: unknown; tenant_id?: string | null } | null;
  if (!body?.market) return err(c, "validation_error", "body.market required", 400);
  try {
    const r = await registerMarket(c.env, cfg, body.market, body.tenant_id ?? null);
    return ok(c, r, 201);
  } catch (e) {
    return err(c, "validation_error", String(e).slice(0, 400), 400);
  }
});

internal.get("/markets/:id", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const client = db(c.env);
  const id = c.req.param("id");
  const [m, w, e, r, l] = await Promise.all([
    client.from("markets").select("*").eq("id", id).single(),
    client.from("watches").select("id, source_kind, source_ref, poll_interval_s, next_poll_at, lease_until, etag, cursor, coverage, last_evidence_hash, last_canonical_hash, last_http_status, last_polled_at, last_error, consecutive_errors, backlog").eq("market_id", id),
    client.from("evidence").select("id, source_kind, source_url, observed_at, claimed_at, fetched_at, http_status, raw_sha256, canonical_sha256, raw_bytes, raw_r2_key, coverage, injection_markers, created_at").eq("market_id", id).order("created_at", { ascending: false }).limit(5),
    client.from("resolutions").select("id, mode, status_row, resolution_status, winning_outcome, confidence_score, error_code, error_reason, caveats, determination_basis, jev_model, thresholds_version, credits_charged, duration_ms, jev_ms, created_at").eq("market_id", id).order("created_at", { ascending: false }).limit(5),
    client.from("loop_runs").select("started_at, outcome, rows_written, duration_ms, error, meta").eq("loop_name", "watch").contains("meta", { market_id: id }).order("started_at", { ascending: false }).limit(5),
  ]);
  if (m.error) return err(c, "not_found", m.error.message, 404);
  return ok(c, { market: m.data, watches: w.data ?? [], evidence: e.data ?? [], resolutions: r.data ?? [], loop_runs: l.data ?? [] });
});

/** Founder-only tenant onboarding for the first 90 days: create a tenant and its first key (raw key shown once). */
internal.post("/tenants", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const b = (await c.req.json().catch(() => ({}))) as { display_name?: string; contact?: string; wallet_address?: string; plan?: string; credits?: number; environment?: "live" | "test"; strict_v0?: boolean; watch_limit?: number };
  if (!b.display_name) return err(c, "validation_error", "display_name required", 400);
  const client = db(c.env);
  const { data: t, error } = await client.from("tenants").insert({ display_name: b.display_name, contact: b.contact ?? null, wallet_address: b.wallet_address ? b.wallet_address.toLowerCase() : null, plan: b.plan ?? "free", strict_v0: !!b.strict_v0, watch_limit: b.watch_limit ?? 5 }).select("id").single();
  if (error || !t) return err(c, "validation_error", error?.message ?? "tenant insert failed", 400);
  if (b.credits && b.credits > 0) await rpc(client, "grant_credits", { p_tenant: t.id, p_amount: b.credits, p_note: "onboarding grant" });
  const key = await mintKey(b.environment ?? "test");
  const { data: k, error: ke } = await client.from("api_keys").insert({ tenant_id: t.id, key_hash: key.hash, key_prefix: key.prefix, name: "initial", environment: b.environment ?? "test", daily_cap: 1000 }).select("id").single();
  if (ke || !k) return err(c, "internal_error", ke?.message ?? "key insert failed", 500);
  return ok(c, { tenant_id: t.id, key_id: k.id, key: key.raw, note: "Shown once." }, 201);
});

internal.post("/reconcile", async (c) => { if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403); return ok(c, await runReconcile(c.env)); });
// Its own invocation and the scan never throws, so it gets every subrequest: the way past a block too big for the cron's share.
internal.post("/deposits/scan", async (c) => { if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403); return ok(c, await scanDeposits(c.env, parseConfig(c.env), new Budget(INVOCATION_SUBREQUESTS))); });
internal.post("/webhooks/drain", async (c) => { if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403); return ok(c, await drainWebhooks(c.env, 10)); });

/** Read the R2 diagnostics the scheduled handler writes when an insert fails. */
internal.get("/diag", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const list = await c.env.BACKUPS.list({ prefix: "diag/", limit: 20 });
  const keys = list.objects.map((o) => o.key).sort().reverse();
  const latest = keys[0] ? await (await c.env.BACKUPS.get(keys[0]))?.text() : null;
  return ok(c, { count: keys.length, keys: keys.slice(0, 10), latest });
});

/** GitHub Actions posts eval summaries here with the scoped insert-only key. */
internal.post("/eval-report", async (c) => {
  const k = bearer(c);
  if (!k || !safeEqual(k, c.env.EVAL_REPORT_KEY)) return err(c, "forbidden", "eval report key required", 403);
  const body = await c.req.json().catch(() => null);
  if (!body || typeof body !== "object") return err(c, "validation_error", "json body required", 400);
  try {
    const id = await rpc<number>(db(c.env), "report_eval_run", { p: body });
    return ok(c, { id }, 201);
  } catch (e) { return err(c, "validation_error", String(e).slice(0, 300), 400); }
});

/** Latency benchmark from this Worker's colo: N fixed 3k-token probes; writes bench_runs. */
/**
 * One-time Telegram wiring check, run from the Worker because api.telegram.org is unreachable from the founder's network.
 * Sets the bot description, confirms the bot can post in the channel, sends one operator DM, and posts + pins the
 * channel disclaimer (only when the channel has no pinned message yet). Idempotent; never prints the token.
 */
internal.post("/bot/setup", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const tok = c.env.TELEGRAM_BOT_TOKEN, ch = c.env.TELEGRAM_CHANNEL_ID, op = c.env.TELEGRAM_OPERATOR_CHAT_ID;
  if (!tok || !ch) return err(c, "config_error", "TELEGRAM_BOT_TOKEN / TELEGRAM_CHANNEL_ID unset", 400);
  const api = async (m: string, body: Record<string, unknown> = {}) => {
    const r = await fetch(`https://api.telegram.org/bot${tok}/${m}`, { method: "POST", headers: { "content-type": "application/json" }, body: JSON.stringify(body), signal: AbortSignal.timeout(8000) });
    return (await r.json()) as { ok: boolean; result?: any; description?: string };
  };
  const out: Record<string, unknown> = {};
  const me = await api("getMe"); out.bot = me.ok ? `@${me.result.username}` : me.description;
  if (!me.ok) return ok(c, out, 502);
  const disc = "Automated shadow settlements for long-tail prediction markets. Every verdict is committed by hash before the official outcome and revealed after. Informational signal, not financial advice, not an oracle of record.";
  out.description_set = (await api("setMyDescription", { description: disc })).ok;
  out.short_description_set = (await api("setMyShortDescription", { short_description: "Commit-reveal shadow settlements for long-tail prediction markets. Not financial advice." })).ok;
  const chat = await api("getChat", { chat_id: ch }); out.channel = chat.ok ? { title: chat.result.title, id: chat.result.id, pinned: chat.result.pinned_message?.message_id ?? null } : chat.description;
  const adm = await api("getChatMember", { chat_id: ch, user_id: me.result.id }); out.bot_membership = adm.ok ? { status: adm.result.status, can_post: adm.result.can_post_messages ?? null } : adm.description;
  if (op) { const dm = await api("sendMessage", { chat_id: op, text: "Resolve operator alerts are wired. One-time test from the settle-bot." }); out.operator_dm = dm.ok ? "sent" : dm.description; }
  if (chat.ok && !chat.result.pinned_message) {
    const text = `Resolve Settlement Feed\n\nThis channel publishes automated shadow settlements for long-tail prediction markets. Each verdict is posted first as a commitment hash before the market's official outcome, then revealed in a reply once the official resolution lands, so every call can be checked after the fact and none can be edited.\n\nInformational signal only. Not financial advice. Not an oracle of record.\n\nTrack record: ${c.env.RESOLVE_PUBLIC_URL ?? ""}/v1/track-record`;
    const post = await api("sendMessage", { chat_id: ch, text, disable_web_page_preview: true });
    out.disclaimer_post = post.ok ? post.result.message_id : post.description;
    if (post.ok) out.pinned = (await api("pinChatMessage", { chat_id: ch, message_id: post.result.message_id, disable_notification: true })).ok;
  }
  return ok(c, out);
});

internal.post("/bench", async (c) => {
  if (!isAdmin(c)) return err(c, "forbidden", "admin key required", 403);
  const cfg = parseConfig(c.env);
  if (!c.env.TYPESAFE_API_KEY) return err(c, "UPSTREAM_UNAVAILABLE", "TYPESAFE_API_KEY not configured", 503);
  const n = Math.min(50, Math.max(3, Number(c.req.query("n") ?? "20")));
  const caller = makeJevCaller({ apiKey: c.env.TYPESAFE_API_KEY, timeoutMs: cfg.jevTimeoutMs * 2 });
  const filler = "The maintainers merged the change after review and the release shipped the same day. ".repeat(120);
  const lat: number[] = [];
  let model = cfg.jevModel;
  for (let i = 0; i < n; i++) {
    const t0 = Date.now();
    try {
      const r = await caller({ model: cfg.jevModel, state: { market: { event_statement: "The pull request is merged", option_a: "Yes", option_b: "No" }, evidence: { source_kind: "web", delimiter: "bench", note: "benchmark", windows: [filler + ` probe ${i}`] } }, questions: { outcome: { type: "choice", instructions: "Which option the evidence establishes", criteria: { OPTION_A: "Yes", OPTION_B: "No", NOT_DETERMINABLE: "Neither" } }, same_subject: { type: "noul", instructions: "Same subject?" } } });
      model = String((r.json as { model?: string }).model ?? model);
      lat.push(Date.now() - t0);
    } catch (e) { return err(c, "UPSTREAM_UNAVAILABLE", `probe ${i} failed: ${String(e).slice(0, 200)}`, 503); }
  }
  const s = [...lat].sort((a, b) => a - b);
  const p50 = s[Math.floor(s.length / 2)]!, p95 = s[Math.min(s.length - 1, Math.floor(s.length * 0.95))]!, max = s[s.length - 1]!;
  const colo = ((c.req.raw as Request & { cf?: { colo?: string } }).cf?.colo) ?? null;
  const { error } = await db(c.env).from("bench_runs").insert({ runner: "cf_worker", colo, n, p50_ms: p50, p95_ms: p95, max_ms: max, jev_model: model, git_sha: cfg.gitSha, meta: { latencies: lat } });
  if (error) return err(c, "internal_error", error.message, 500);
  return ok(c, { runner: "cf_worker", colo, n, p50_ms: p50, p95_ms: p95, max_ms: max, model });
});